requirements.txt
Pyrebase4==4.8.0
pillow==11.0.0
tensorflow==2.18.0
